pacify MSVC

+++ arangod/RocksDBEngine/RocksDBEngine.cpp
@@ -439,7 +439,7 @@ void RocksDBEngine::start() {
   _options.use_direct_io_for_flush_and_compaction = opts._useDirectIoForFlushAndCompaction;
 
   _options.target_file_size_base = opts._targetFileSizeBase;
-  _options.target_file_size_multiplier = opts._targetFileSizeMultiplier;
+  _options.target_file_size_multiplier = static_cast<int>(opts._targetFileSizeMultiplier);
   // during startup, limit the total WAL size to a small value so we do not see
   // large WAL files created at startup.
   // Instead, we will start with a small value here and up it later in the startup process
